Improvements after review
