+++ .eslintrc.js
@@ -34,5 +34,6 @@ module.exports = {
       },
     ],
     '@typescript-eslint/no-unused-vars': ['error', { argsIgnorePattern: '^_' }],
+    '@typescript-eslint/no-var-requires': 0,
   },
 };
