--- Example/jest-setup.js
@@ -1,5 +1,0 @@
-require('../src/reanimated2/Jest').SetUpTests();
-// jest.mock('react-native-reanimated', () => {
-//     const Reanimated = require('../src/reanimated2')
-//     return Reanimated;
-// });
