+++ __tests__/SharedValue.test.js
@@ -1,14 +1,14 @@
 import React from 'react';
 import { View, Button } from 'react-native';
 import { render, fireEvent } from '@testing-library/react-native';
-import { useSharedValue } from 'react-native-reanimated';
+import { useSharedValue } from '../src';
 
-const TestComponent1 = () => {
+const ViewTestComponent = () => {
   const sv = useSharedValue(1);
   return <View testID="view">{sv.value}</View>;
 };
 
-function TestComponent2() {
+function TestComponentWithSV() {
   const sv = useSharedValue(1);
 
   return (
@@ -27,13 +27,13 @@ function TestComponent2() {
 
 describe('SharedValue basic render', () => {
   test('contains default value', () => {
-    const { getByTestId } = render(<TestComponent1 />);
+    const { getByTestId } = render(<ViewTestComponent />);
     const view = getByTestId('view');
     expect(view.children[0]).toBe('1');
   });
 
   test('no changes value after click', () => {
-    const { getByTestId } = render(<TestComponent2 />);
+    const { getByTestId } = render(<TestComponentWithSV />);
     const view = getByTestId('view');
     const button = getByTestId('button');
