+++ jest-setup.js
@@ -1,0 +1,1 @@
+require('./src/reanimated2/jestUtils').setUpTests();
