+++ jest.config.js
@@ -1,4 +1,6 @@
 module.exports = {
   preset: 'react-native',
   modulePathIgnorePatterns: ['Example', 'docs', 'lib'],
+  setupFiles: ['./jest-setup.js'],
+  setupFilesAfterEnv: ['@testing-library/jest-native/extend-expect'],
 };
