+++ package.json
@@ -51,9 +51,7 @@
   "homepage": "https://github.com/software-mansion/react-native-reanimated#readme",
   "dependencies": {
     "@babel/plugin-transform-object-assign": "^7.10.4",
-    "expect": "^26.6.2",
     "fbjs": "^3.0.0",
-    "mockdate": "^3.0.2",
     "string-hash-64": "^1.0.3"
   },
   "peerDependencies": {
@@ -68,6 +66,9 @@
     "@babel/preset-typescript": "^7.7.4",
     "@react-native-community/bob": "^0.14.3",
     "@react-native-community/eslint-config": "^0.0.5",
+    "@testing-library/jest-native": "^3.4.3",
+    "@testing-library/react-hooks": "^5.0.3",
+    "@testing-library/react-native": "^7.1.0",
     "@types/babel-types": "^7.0.9",
     "@types/babel__core": "^7.1.12",
     "@types/babel__generator": "^7.6.2",
@@ -87,9 +88,11 @@
     "eslint-plugin-node": "^10.0.0",
     "eslint-plugin-promise": "^4.2.1",
     "eslint-plugin-standard": "^4.0.1",
+    "expect": "^26.6.2",
     "husky": "^4.2.5",
     "jest": "^24.9.0",
     "lint-staged": "^10.2.11",
+    "mockdate": "^3.0.2",
     "prettier": "^2.2.1",
     "react": "17.0.1",
     "react-native": "0.64.0-rc.2",
