+++ src/ReanimatedModule.native.js
@@ -1,5 +1,5 @@
 let Module;
-if(process.env.JEST_WORKER_ID) {
+if (process.env.JEST_WORKER_ID) {
   Module = require('./ReanimatedModuleCompat').default;
 } else {
   Module = require('react-native').NativeModules.ReanimatedModule;
