+++ src/reanimated2/Hooks.js
@@ -322,7 +322,7 @@ export function useAnimatedStyle(updater, dependencies, adapters) {
   const viewDescriptor = useSharedValue({ tag: -1, name: null }, false);
   const animatedStyle = useRef({});
   const initRef = useRef(null);
-  const inputs = Object.values(updater._closure);
+  const inputs = Object.values(updater._closure ? updater._closure : {});
   const viewRef = useRef(null);
   adapters = !adapters || Array.isArray(adapters) ? adapters : [adapters];
   const adaptersHash = adapters ? buildWorkletsHash(adapters) : null;
