+++ src/reanimated2/NativeReanimated.js
@@ -1,7 +1,7 @@
 let reanimatedJS;
-if(process.env.JEST_WORKER_ID) {
-  reanimatedJS = require('./js-reanimated/index.web').default
+if (process.env.JEST_WORKER_ID) {
+  reanimatedJS = require('./js-reanimated/index.web').default;
 } else {
-  reanimatedJS = require('./js-reanimated/index')
+  reanimatedJS = require('./js-reanimated/index');
 }
 export default reanimatedJS;
